+++ tests/testthat/test-get_naturalearth.R
@@ -68,3 +68,32 @@ test_that("get_naturalearth works with non-default scale", {
 })
 
 
+test_that("get_naturalearth() works at edge of WGS84", {
+
+  # Construct a psuedo BirdFlow object that has a crs centered on the edge
+  # of the wgs84 projection (used by rnaturalearth)
+  seam_crs <- crs("+proj=moll +lon_0=180 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs +type=crs")  # mollweide centered on 180 deg lon.
+  bf <- new_BirdFlow()
+  bf$geom$crs <- seam_crs
+  bf$geom$ext <- c(-2000000, 2000000,-5000000, 1000000)
+
+  # Visualize what we are hoping to do
+  if(FALSE){ # This code is to run manually while debugging
+    all_coast <- rnaturalearth::ne_coastline(returnclass = "sf")
+    coast_proj <- sf::st_transform(all_coast, seam_crs)
+    op <- par( no.readonly = TRUE)
+    par(oma = rep(0, 4), mar = rep(0, 4))
+    plot(coast_proj[, "geometry"])
+    abline(v = c(-2000000, 2000000), col = "blue")
+    abline(h = c(-5000000, 1000000), col = "red")
+
+    par(op)
+  }  # end skip visualization
+
+  expect_no_error( coast <- get_coastline(bf, scale = 110) )
+  expect_snapshot(coast) # eastern Australia and islands
+
+})
+
+
+